requirements.txt
Appium-Python-Client==4.4.0
pytest==8.3.4
pytest-html==4.1.1
allure-pytest==2.13.5
selenium==4.27.1
jsonschema==4.23.0
pytest-xdist==3.6.1
pytest-bdd==8.1.0
python-dotenv==1.0.1
